Move mock struct to export_test
