+++ application/service/runner/runner_test.go
@@ -14,7 +14,6 @@ import (
 	"github.com/google/uuid"
 	"github.com/pkg/errors"
 	"gopkg.in/src-d/go-git.v4/plumbing"
-	"io"
 	"net/url"
 	"os"
 	"path"
@@ -59,14 +58,14 @@ func TestRunnerImpl_Run_Normal(t *testing.T) {
 			mockDocker.EXPECT().
 				Build(gomock.Any(), gomock.Any(), gomock.Any(), gomock.Eq("./Dockerfile")).
 				Times(1).
-				Return(&MockBuildLog{}, nil)
+				Return(&runner.MockBuildLog{}, nil)
 			mockDocker.EXPECT().
 				Build(gomock.Any(), gomock.Any(), gomock.Any(), gomock.Not("./Dockerfile")).
 				Return(nil, errors.New("must not call this"))
 			mockDocker.EXPECT().
 				Run(gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any()).
 				Times(1).
-				Return("", &MockJobLog{}, nil)
+				Return("", &runner.MockJobLog{}, nil)
 			mockDocker.EXPECT().
 				ExitCode(gomock.Any(), gomock.Any()).
 				AnyTimes().
@@ -100,7 +99,7 @@ func TestRunnerImpl_Run_Normal(t *testing.T) {
 			}
 
 			// and
-			repo := &MockRepo{FullName: "duck8823/duci", SSHURL: "git@github.com:duck8823/duci.git"}
+			repo := &runner.MockRepo{FullName: "duck8823/duci", SSHURL: "git@github.com:duck8823/duci.git"}
 
 			// when
 			err := r.Run(
@@ -146,14 +145,14 @@ func TestRunnerImpl_Run_Normal(t *testing.T) {
 			mockDocker := mock_docker.NewMockClient(ctrl)
 			mockDocker.EXPECT().
 				Build(gomock.Any(), gomock.Any(), gomock.Any(), gomock.Eq(".duci/Dockerfile")).
-				Return(&MockBuildLog{}, nil)
+				Return(&runner.MockBuildLog{}, nil)
 			mockDocker.EXPECT().
 				Build(gomock.Any(), gomock.Any(), gomock.Any(), gomock.Not(".duci/Dockerfile")).
 				Return(nil, errors.New("must not call this"))
 			mockDocker.EXPECT().
 				Run(gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any()).
 				Times(1).
-				Return("", &MockJobLog{}, nil)
+				Return("", &runner.MockJobLog{}, nil)
 			mockDocker.EXPECT().
 				ExitCode(gomock.Any(), gomock.Any()).
 				AnyTimes().
@@ -187,7 +186,7 @@ func TestRunnerImpl_Run_Normal(t *testing.T) {
 			}
 
 			// and
-			repo := &MockRepo{FullName: "duck8823/duci", SSHURL: "git@github.com:duck8823/duci.git"}
+			repo := &runner.MockRepo{FullName: "duck8823/duci", SSHURL: "git@github.com:duck8823/duci.git"}
 
 			// when
 			err := r.Run(
@@ -234,11 +233,11 @@ func TestRunnerImpl_Run_Normal(t *testing.T) {
 		mockDocker := mock_docker.NewMockClient(ctrl)
 		mockDocker.EXPECT().
 			Build(gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any()).
-			Return(&MockBuildLog{}, nil)
+			Return(&runner.MockBuildLog{}, nil)
 		mockDocker.EXPECT().
 			Run(gomock.Any(), gomock.Eq(docker.RuntimeOptions{Volumes: []string{"/hello:/hello"}}), gomock.Any(), gomock.Any()).
 			Times(1).
-			Return("", &MockJobLog{}, nil)
+			Return("", &runner.MockJobLog{}, nil)
 		mockDocker.EXPECT().
 			Run(gomock.Any(), gomock.Not(docker.RuntimeOptions{Volumes: []string{"/hello:/hello"}}), gomock.Any(), gomock.Any()).
 			Return("", nil, errors.New("must not call this"))
@@ -275,7 +274,7 @@ func TestRunnerImpl_Run_Normal(t *testing.T) {
 		}
 
 		// and
-		repo := &MockRepo{FullName: "duck8823/duci", SSHURL: "git@github.com:duck8823/duci.git"}
+		repo := &runner.MockRepo{FullName: "duck8823/duci", SSHURL: "git@github.com:duck8823/duci.git"}
 
 		// when
 		err := r.Run(
@@ -349,7 +348,7 @@ func TestRunnerImpl_Run_NonNormal(t *testing.T) {
 		}
 
 		// and
-		repo := &MockRepo{FullName: "duck8823/duci", SSHURL: "git@github.com:duck8823/duci.git"}
+		repo := &runner.MockRepo{FullName: "duck8823/duci", SSHURL: "git@github.com:duck8823/duci.git"}
 
 		// when
 		err := r.Run(
@@ -385,7 +384,7 @@ func TestRunnerImpl_Run_NonNormal(t *testing.T) {
 		}
 
 		// and
-		repo := &MockRepo{FullName: "duck8823/duci", SSHURL: "git@github.com:duck8823/duci.git"}
+		repo := &runner.MockRepo{FullName: "duck8823/duci", SSHURL: "git@github.com:duck8823/duci.git"}
 
 		// when
 		err := r.Run(
@@ -454,7 +453,7 @@ func TestRunnerImpl_Run_NonNormal(t *testing.T) {
 		}
 
 		// and
-		repo := &MockRepo{FullName: "duck8823/duci", SSHURL: "git@github.com:duck8823/duci.git"}
+		repo := &runner.MockRepo{FullName: "duck8823/duci", SSHURL: "git@github.com:duck8823/duci.git"}
 
 		// when
 		err := r.Run(
@@ -524,7 +523,7 @@ func TestRunnerImpl_Run_NonNormal(t *testing.T) {
 		}
 
 		// and
-		repo := &MockRepo{FullName: "duck8823/duci", SSHURL: "git@github.com:duck8823/duci.git"}
+		repo := &runner.MockRepo{FullName: "duck8823/duci", SSHURL: "git@github.com:duck8823/duci.git"}
 
 		// when
 		err := r.Run(
@@ -557,7 +556,7 @@ func TestRunnerImpl_Run_NonNormal(t *testing.T) {
 		mockDocker.EXPECT().
 			Build(gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any()).
 			Times(1).
-			Return(&MockBuildLog{}, nil)
+			Return(&runner.MockBuildLog{}, nil)
 		mockDocker.EXPECT().
 			Run(gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any()).
 			Times(1).
@@ -595,7 +594,7 @@ func TestRunnerImpl_Run_NonNormal(t *testing.T) {
 		}
 
 		// and
-		repo := &MockRepo{FullName: "duck8823/duci", SSHURL: "git@github.com:duck8823/duci.git"}
+		repo := &runner.MockRepo{FullName: "duck8823/duci", SSHURL: "git@github.com:duck8823/duci.git"}
 
 		// when
 		err := r.Run(
@@ -630,11 +629,11 @@ func TestRunnerImpl_Run_NonNormal(t *testing.T) {
 		mockDocker := mock_docker.NewMockClient(ctrl)
 		mockDocker.EXPECT().
 			Build(gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any()).
-			Return(&MockBuildLog{}, nil)
+			Return(&runner.MockBuildLog{}, nil)
 		mockDocker.EXPECT().
 			Run(gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any()).
 			Times(1).
-			Return("", &MockJobLog{}, nil)
+			Return("", &runner.MockJobLog{}, nil)
 		mockDocker.EXPECT().
 			ExitCode(gomock.Any(), gomock.Any()).
 			AnyTimes().
@@ -668,7 +667,7 @@ func TestRunnerImpl_Run_NonNormal(t *testing.T) {
 		}
 
 		// and
-		repo := &MockRepo{FullName: "duck8823/duci", SSHURL: "git@github.com:duck8823/duci.git"}
+		repo := &runner.MockRepo{FullName: "duck8823/duci", SSHURL: "git@github.com:duck8823/duci.git"}
 
 		// when
 		err := r.Run(
@@ -701,11 +700,11 @@ func TestRunnerImpl_Run_NonNormal(t *testing.T) {
 		mockDocker.EXPECT().
 			Build(gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any()).
 			Times(1).
-			Return(&MockBuildLog{}, nil)
+			Return(&runner.MockBuildLog{}, nil)
 		mockDocker.EXPECT().
 			Run(gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any()).
 			Times(1).
-			Return("", &MockJobLog{}, nil)
+			Return("", &runner.MockJobLog{}, nil)
 		mockDocker.EXPECT().
 			ExitCode(gomock.Any(), gomock.Any()).
 			AnyTimes().
@@ -739,7 +738,7 @@ func TestRunnerImpl_Run_NonNormal(t *testing.T) {
 		}
 
 		// and
-		repo := &MockRepo{FullName: "duck8823/duci", SSHURL: "git@github.com:duck8823/duci.git"}
+		repo := &runner.MockRepo{FullName: "duck8823/duci", SSHURL: "git@github.com:duck8823/duci.git"}
 
 		// when
 		err := r.Run(
@@ -774,13 +773,13 @@ func TestRunnerImpl_Run_NonNormal(t *testing.T) {
 		mockDocker.EXPECT().
 			Build(gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any()).
 			Times(1).
-			Return(&MockBuildLog{}, nil)
+			Return(&runner.MockBuildLog{}, nil)
 		mockDocker.EXPECT().
 			Run(gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any()).
 			Times(1).
 			DoAndReturn(func(ctx context.Context, opts docker.RuntimeOptions, tag string, cmd ...string) (string, docker.Log, error) {
 				time.Sleep(3 * time.Second)
-				return "container_id", &MockJobLog{}, nil
+				return "container_id", &runner.MockJobLog{}, nil
 			})
 		mockDocker.EXPECT().
 			ExitCode(gomock.Any(), gomock.Any()).
@@ -815,7 +814,7 @@ func TestRunnerImpl_Run_NonNormal(t *testing.T) {
 		}
 
 		// and
-		repo := &MockRepo{FullName: "duck8823/duci", SSHURL: "git@github.com:duck8823/duci.git"}
+		repo := &runner.MockRepo{FullName: "duck8823/duci", SSHURL: "git@github.com:duck8823/duci.git"}
 
 		// when
 		err := r.Run(
@@ -829,38 +828,6 @@ func TestRunnerImpl_Run_NonNormal(t *testing.T) {
 			t.Errorf("error must be runner.ErrFailure, but got %+v", err)
 		}
 	})
-}
-
-type MockRepo struct {
-	FullName string
-	SSHURL   string
-	CloneURL string
-}
-
-func (r *MockRepo) GetFullName() string {
-	return r.FullName
-}
-
-func (r *MockRepo) GetSSHURL() string {
-	return r.SSHURL
-}
-
-func (r *MockRepo) GetCloneURL() string {
-	return r.CloneURL
-}
-
-type MockBuildLog struct {
-}
-
-func (l *MockBuildLog) ReadLine() (*docker.LogLine, error) {
-	return &docker.LogLine{Timestamp: time.Now(), Message: []byte("{\"stream\":\"Hello World,\"}")}, io.EOF
-}
-
-type MockJobLog struct {
-}
-
-func (l *MockJobLog) ReadLine() (*docker.LogLine, error) {
-	return &docker.LogLine{Timestamp: time.Now(), Message: []byte("Hello World,")}, io.EOF
 }
 
 func cloneSuccess(_ interface{}, dir string, _ interface{}) error {
